remove unneeded mobile footer import statements in mobile pages
