+++ src/components/MobileFooter.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import '../styles/MobileFooter.css';
 
-const MobileFooter = (props) => {
+const MobileFooter2 = (props) => {
   const style = {
     backgroundColor: '' + props.color + '',
   };
@@ -9,8 +9,8 @@ const MobileFooter = (props) => {
   return (
     <Fragment>
       <div className='connect-container' style={style}>
-        <div className='connect-fb'>
-          <div className='connect-fb-icon'>
+        <div className='mobile-social-icons-container'>
+          <div className='mobile-icon' id='facebook-icon'>
             <a href='https://www.facebook.com/yhwhapparel'>
               <img
                 src={require('../images/facebook_icon.png')}
@@ -18,17 +18,7 @@ const MobileFooter = (props) => {
               ></img>
             </a>
           </div>
-          <div className='connect-fb-caption-container'>
-            <div className='connect-fb-title'>
-              <h4>FOLLOW AND LIKE US ON FACEBOOK</h4>
-            </div>
-            <div className='connect-fb-caption'>
-              <h5>Get notified about our events and campus sales.</h5>
-            </div>
-          </div>
-        </div>
-        <div className='connect-ig'>
-          <div className='connect-ig-icon'>
+          <div className='mobile-icon' id='instagram-icon'>
             <a href='https://www.instagram.com/yhwhapparel'>
               <img
                 src={require('../images/instagram_icon.png')}
@@ -36,31 +26,13 @@ const MobileFooter = (props) => {
               ></img>
             </a>
           </div>
-          <div className='connect-ig-caption-container'>
-            <div className='connect-ig-title'>
-              <h4>FOLLOW US ON INSTAGRAM</h4>
-            </div>
-            <div className='connect-ig-caption'>
-              <h5>Stay updated on latest drops.</h5>
-            </div>
-          </div>
-        </div>
-        <div className='connect-medium'>
-          <div className='connect-medium-icon'>
+          <div className='mobile-icon' id='footer-icon'>
             <a href='https://medium.com/@yhwhapparel/'>
               <img
                 src={require('../images/medium_icon.png')}
                 alt='medium-icon'
               ></img>
             </a>
-          </div>
-          <div className='connect-medium-caption-container'>
-            <div className='connect-medium-title'>
-              <h4>FOLLOW US ON MEDIUM</h4>
-            </div>
-            <div className='connect-medium-caption'>
-              <h5>Keep up with our company and follow us on our journey.</h5>
-            </div>
           </div>
         </div>
       </div>
@@ -88,4 +60,4 @@ const MobileFooter = (props) => {
   );
 };
 
-export default MobileFooter;
+export default MobileFooter2;
